+++ jsx_stream/jsx_runtime.ts
@@ -1,9 +1,12 @@
 // deno-lint-ignore-file no-explicit-any
 import { streamComponent } from "./stream_component.ts";
 import { streamElement } from "./stream_element.ts";
-import { streamFragment } from "./stream_fragment.ts";
+import { streamFragment } from "@http/html-stream/stream-fragment";
 import { streamUnknown } from "./stream_unknown.ts";
-import type { HtmlNode } from "@http/html-stream/types";
+import { html } from "@http/html-stream/template";
+import { escape, partialHtml, safe } from "@http/html-stream/token";
+import { isValidAttr } from "@http/html-stream/util";
+import type { AttrName, HtmlNode } from "@http/html-stream/types";
 
 export function jsx(type: any, props: any): HtmlNode {
   if (typeof type === "function") {
@@ -14,6 +17,18 @@ export function jsx(type: any, props: any): HtmlNode {
     return streamElement(type, props);
   } else {
     return streamUnknown(type);
+  }
+}
+
+export const jsxTemplate = html;
+
+export const jsxEscape = streamFragment;
+
+export function jsxAttr(name: AttrName, value: unknown): string {
+  if (isValidAttr(name, value)) {
+    return partialHtml(`${name}="${escape(String(value))}"`);
+  } else {
+    return safe("");
   }
 }